runProgram() now uses spawn()
