+++ lib/common.js
@@ -1,18 +1,19 @@
-const { execFile } = require('child_process');
+const { spawn } = require('child_process');
+const byline = require('byline');
 
-function runProgram(command, args, dir) {
-  const options = {
-    maxBuffer: 10 * 1024 * 1024,
-    cwd: dir || undefined,
-  };
+// function runProgram(command, args, dir) {
+//   const options = {
+//     maxBuffer: 10 * 1024 * 1024,
+//     cwd: dir || undefined,
+//   };
 
-  return new Promise(resolve =>
-    execFile(command, args, options, (err, stdout, stderr) => {
-      const description = command + ' ' + args.map(x => `"${x}"`).join(' ');
-      resolve({ exitCode: err ? err.code : 0, stdout, stderr, description });
-    })
-  );
-}
+//   return new Promise(resolve =>
+//     execFile(command, args, options, (err, stdout, stderr) => {
+//       const description = command + ' ' + args.map(x => `"${x}"`).join(' ');
+//       resolve({ exitCode: err ? err.code : 0, stdout, stderr, description });
+//     })
+//   );
+// }
 
 function reportAudit(summary, config, parsedOutput) {
   const { whitelist } = config;
@@ -47,6 +48,35 @@ function reportAudit(summary, config, parsedOutput) {
     throw new Error(err);
   }
   return summary;
+}
+
+function listenTo(stream, reject, listener) {
+  stream.setEncoding('utf8');
+  const linedStream = byline.createStream(stream);
+  linedStream.on('readable', () => {
+    while (true) {
+      let line = linedStream.read();
+      if (line === null) {
+        break;
+      }
+
+      try {
+        listener(line);
+      } catch (e) {
+        reject(e);
+      }
+    }
+  });
+}
+
+function runProgram(command, args, options, stdoutListener, stderrListener) {
+  return new Promise((resolve, reject) => {
+    const proc = spawn(command, args, options);
+
+    listenTo(proc.stdout, reject, stdoutListener);
+    listenTo(proc.stderr, reject, stderrListener);
+    proc.on('close', () => resolve());
+  });
 }
 
 module.exports = {
