+++ lib/npm-auditer.js
@@ -7,15 +7,34 @@ const { runProgram, reportAudit } = require('./common');
 const Model = require('./Model');
 
 function runNpmAudit(dir) {
+  const stdoutBuffer = [];
+  function stdoutListener(line) {
+    stdoutBuffer.push(line);
+  }
+
+  const stderrBuffer = [];
+  function stderrListener(line) {
+    stderrBuffer.push(line);
+  }
+
   return Promise.resolve()
-    .then(() => runProgram('npm', ['audit', '--json'], dir))
-    .then(({ exitCode, stdout, stderr, description }) => {
-      if (stderr) {
+    .then(() =>
+      runProgram(
+        'npm',
+        ['audit', '--json'],
+        { cwd: dir },
+        stdoutListener,
+        stderrListener
+      )
+    )
+    .then(() => {
+      if (stderrBuffer.length) {
         throw new Error(
-          `Invocation of ${description} failed (exit code: ${exitCode}):\n${stderr}`
+          `Invocation of npm audit failed:\n${stderrBuffer.join('\n')}`
         );
       }
 
+      const stdout = stdoutBuffer.join('\n');
       try {
         return JSON.parse(stdout);
       } catch (e) {
