+++ lib/yarn-auditer.js
@@ -3,12 +3,10 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-const { reportAudit } = require('./common');
+const { reportAudit, runProgram } = require('./common');
 const childProcess = require('child_process');
-const spawn = require('cross-spawn');
 const semver = require('semver');
 const Model = require('./Model');
-const byline = require('byline');
 
 const MINIMUM_YARN_VERSION = '1.12.3';
 
@@ -22,35 +20,6 @@ function getYarnVersion() {
 
 function yarnSupportsAudit(yarnVersion) {
   return semver.gte(yarnVersion, MINIMUM_YARN_VERSION);
-}
-
-function listenTo(stream, reject, listener) {
-  stream.setEncoding('utf8');
-  const linedStream = byline.createStream(stream);
-  linedStream.on('readable', () => {
-    while (true) {
-      let line = linedStream.read();
-      if (line === null) {
-        break;
-      }
-
-      try {
-        listener(line);
-      } catch (e) {
-        reject(e);
-      }
-    }
-  });
-}
-
-function run(command, args, options, stdoutListener, stderrListener) {
-  return new Promise((resolve, reject) => {
-    const proc = spawn(command, args, options);
-
-    listenTo(proc.stdout, reject, stdoutListener);
-    listenTo(proc.stderr, reject, stderrListener);
-    proc.on('close', () => resolve());
-  });
 }
 
 /**
@@ -113,19 +82,23 @@ function audit(config, reporter = reportAudit) {
     }
     const options = { cwd: config.dir };
     const args = ['audit', '--json'];
-    return run('yarn', args, options, stdoutListener, stderrListener).then(
-      () => {
-        if (missingLockFile) {
-          console.warn(
-            '\x1b[33m%s\x1b[0m',
-            'No yarn.lock file. This does not affect auditing, but it may be a mistake.'
-          );
-        }
-
-        const summary = model.getSummary(a => a.id);
-        return reporter(summary);
+    return runProgram(
+      'yarn',
+      args,
+      options,
+      stdoutListener,
+      stderrListener
+    ).then(() => {
+      if (missingLockFile) {
+        console.warn(
+          '\x1b[33m%s\x1b[0m',
+          'No yarn.lock file. This does not affect auditing, but it may be a mistake.'
+        );
       }
-    );
+
+      const summary = model.getSummary(a => a.id);
+      return reporter(summary);
+    });
   });
 }
